mv Loggly test to test_logs module
