+++ tests/test_logs.py
@@ -2,8 +2,10 @@ import pytest
 import os
 import sys
 import requests
+import logging
 
 from reactors.runtime import Reactor
+from reactors.logtypes.loggly_futures_session import LogglyHandler
 
 
 @pytest.fixture(params=['R_tp_opt', 'R_bare'])
@@ -126,3 +128,32 @@ def test_log_redact_inited(R, caplog, capsys, monkeypatch):
     out, err = capsys.readouterr()
     assert message not in err
     assert message in caplog.text
+
+
+@pytest.mark.loggly_auth
+def test_loggly_handler(R_tp_opt, loggly_token, monkeypatch):
+    """Reading Loggly customer token from env var, can POST to Loggly using the
+    LogglyHandler.
+    """
+    monkeypatch.setenv('_REACTOR_LOGGLY_CUSTOMER_TOKEN', loggly_token)
+    r = R_tp_opt()
+
+    # get the LogglyHandler instance
+    assert isinstance(r.loggers.loggly, logging.Logger), type(r.loggers.loggly)
+    lhs = [handler for handler in r.loggers.loggly.handlers 
+           if isinstance(handler, LogglyHandler)]
+    assert len(lhs) == 1
+    lh = lhs[0]
+    assert isinstance(lh, LogglyHandler)
+
+    # try posting and check the status code
+    assert not hasattr(lh, '_resp')
+    r.loggers.loggly.info('hello from LogglyHandler')
+    assert hasattr(lh, '_resp')
+    assert lh._resp.result().status_code == 200
+
+    # now bork the token and check that the response fails to POST
+    bogus_token = "bogus_token"
+    lh.url = 'https://logs-01.loggly.com/inputs/{}/tag/python'.format(bogus_token)
+    r.loggers.loggly.info('this message from LogglyHandler wont go through')
+    assert lh._resp.result().status_code != 200
